backend/requirements.txt
fastapi
uvicorn
google-generativeai
sentence-transformers
faiss-cpu
python-multipart
python-dotenv